Rename to Chart.jsx
