+++ app/components/visualizer/main-viz/templates/Chart.jsx
